pyproject.toml
[tool.poetry]
name = "madlibs_api"
version = "1.0"
description = "REST API for random Madlibs stories."
authors = ["Ray Luna <leon.luna.ray@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.4"
django-cors-headers = "^4.2.0"
django-restframework = "^0.0.1"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
